chore: split codes chunk for max line of 300 lint
